+++ changelog.md
@@ -2,4 +2,6 @@ CHANGELOG
 =========
 
 - r2.0: adds Litecoin/regtest support;
-- r2.1: adds BPcoin main/testnet support;
+- r2.1: 
+  * adds BPcoin main/testnet support;
+  * adds coin::HASH_ALGO field, do not check genesis block if algo not supported;